Move parsing unit tests to be colocated with the parser package
The remaining tests in test/exclusion_test.go, test/intersection_test.go,
and test/intersection_ttu_union_test.go all exercise parser.ParseSchemaString()
and inspect schema struct fields. They belong in pkg/parser/ alongside the
code they test, not in the top-level integration test directory.

- Move 4 parsing tests to pkg/parser/parsing_patterns_test.go as external
  tests (package parser_test) since they use the public API
- Delete the now-empty source files from test/
